+++ herfaty/lib/oldAddProduct.dart
@@ -488,7 +488,7 @@ class _AddProductState extends State<AddProduct> {
                       uploadImageUrl != "") {
                     String prodName = nameController.text;
                     String desc = descController.text;
-                    int amount = int.parse(amountController.text);
+                    double amount = double.parse(amountController.text);
                     double priceN = double.parse(priceController.text);
 
                     /*
